+++ src/main/java/ru/mipt/bit/platformer/GameDesktopLauncher.java
@@ -6,11 +6,10 @@ import com.badlogic.gdx.backends.lwjgl3.Lwjgl3Application;
 import com.badlogic.gdx.backends.lwjgl3.Lwjgl3ApplicationConfiguration;
 import com.badlogic.gdx.graphics.g2d.Batch;
 import com.badlogic.gdx.graphics.g2d.SpriteBatch;
-import com.badlogic.gdx.maps.tiled.TiledMapTileLayer;
 import com.badlogic.gdx.math.GridPoint2;
 import com.badlogic.gdx.math.Interpolation;
 import com.badlogic.gdx.utils.Disposable;
-import input.KeyboardListener;
+import ru.mipt.bit.platformer.input.KeyboardListener;
 import ru.mipt.bit.platformer.roles.Tank;
 import ru.mipt.bit.platformer.roles.Tree;
 import ru.mipt.bit.platformer.ui.MovableTexturedItem;
