+++ src/main/java/ru/mipt/bit/platformer/input/LeftAction.java
@@ -1,24 +1,24 @@
-package input;
+package ru.mipt.bit.platformer.input;
 
 import ru.mipt.bit.platformer.ui.MovableTexturedItem;
 import ru.mipt.bit.platformer.ui.TexturedItem;
 
 import java.util.List;
 
-public class UpAction implements Action {
+public class LeftAction implements Action {
     @Override
     public void accept(MovableTexturedItem player, List<TexturedItem> obstacles) {
         if (!player.progressCompleted()) {
             return;
         }
 
-        player.rotate(90f);
+        player.rotate(-180f);
 
-        if (obstacles.stream().anyMatch(item -> player.hasCommonBorderOnTopWith(item))) {
+        if (obstacles.stream().anyMatch(item -> player.hasCommonBorderOnLeftWith(item))) {
             return;
         }
 
-        player.moveUp();
+        player.moveLeft();
         player.resetProgress();
 
     }
