+++ src/main/java/ru/mipt/bit/platformer/input/RightAction.java
@@ -1,24 +1,25 @@
-package input;
+package ru.mipt.bit.platformer.input;
 
 import ru.mipt.bit.platformer.ui.MovableTexturedItem;
 import ru.mipt.bit.platformer.ui.TexturedItem;
 
 import java.util.List;
 
-public class DownAction implements Action {
+public class RightAction implements Action {
     @Override
     public void accept(MovableTexturedItem player, List<TexturedItem> obstacles) {
         if (!player.progressCompleted()) {
             return;
         }
 
-        player.rotate(-90f);
+        player.rotate(0f);
 
-        if (obstacles.stream().anyMatch(item -> player.hasCommonBorderOnBottomWith(item))) {
+        if (obstacles.stream().anyMatch(item -> player.hasCommonBorderOnRightWith(item))) {
             return;
         }
 
-        player.moveDown();
+        player.moveRight();
         player.resetProgress();
+
     }
 }
